+++ lib/src/zigbee/zigbee.freezed.dart
@@ -20,9 +20,9 @@ mixin _$ZigbeeDevice {
  String get ieeeAddress;/// [manufacturer] is the manufacturer name of the device.
  String? get manufacturer;/// [model] is the model name of the device.
  String? get model;/// [lastSeen] is the date and time when the device was last seen.
- DateTime? get lastSeen;/// [bindedDeviceId] is the ID of the device that is binded to this device.
+@TimestampOrNullConverter() DateTime? get lastSeen;/// [bindedDeviceId] is the ID of the device that is binded to this device.
  String? get bindedDeviceId;/// [zoneId] is the ID of the zone that the device belongs to.
- String get zoneId;/// [exposes] is the list of exposes of the device.
+ String? get zoneId;/// [exposes] is the list of exposes of the device.
  List<ZigbeeDeviceExpose>? get exposes;
 /// Create a copy of ZigbeeDevice
 /// with the given fields replaced by the non-null parameter values.
@@ -56,7 +56,7 @@ abstract mixin class $ZigbeeDeviceCopyWith<$Res>  {
   factory $ZigbeeDeviceCopyWith(ZigbeeDevice value, $Res Function(ZigbeeDevice) _then) = _$ZigbeeDeviceCopyWithImpl;
 @useResult
 $Res call({
- String id, String ieeeAddress, String? manufacturer, String? model, DateTime? lastSeen, String? bindedDeviceId, String zoneId, List<ZigbeeDeviceExpose>? exposes
+ String id, String ieeeAddress, String? manufacturer, String? model,@TimestampOrNullConverter() DateTime? lastSeen, String? bindedDeviceId, String? zoneId, List<ZigbeeDeviceExpose>? exposes
 });
 
 
@@ -73,7 +73,7 @@ class _$ZigbeeDeviceCopyWithImpl<$Res>
 
 /// Create a copy of ZigbeeDevice
 /// with the given fields replaced by the non-null parameter values.
-@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? ieeeAddress = null,Object? manufacturer = freezed,Object? model = freezed,Object? lastSeen = freezed,Object? bindedDeviceId = freezed,Object? zoneId = null,Object? exposes = freezed,}) {
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? ieeeAddress = null,Object? manufacturer = freezed,Object? model = freezed,Object? lastSeen = freezed,Object? bindedDeviceId = freezed,Object? zoneId = freezed,Object? exposes = freezed,}) {
   return _then(_self.copyWith(
 id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
 as String,ieeeAddress: null == ieeeAddress ? _self.ieeeAddress : ieeeAddress // ignore: cast_nullable_to_non_nullable
@@ -81,8 +81,8 @@ as String,manufacturer: freezed == manufacturer ? _self.manufacturer : manufactu
 as String?,model: freezed == model ? _self.model : model // ignore: cast_nullable_to_non_nullable
 as String?,lastSeen: freezed == lastSeen ? _self.lastSeen : lastSeen // ignore: cast_nullable_to_non_nullable
 as DateTime?,bindedDeviceId: freezed == bindedDeviceId ? _self.bindedDeviceId : bindedDeviceId // ignore: cast_nullable_to_non_nullable
-as String?,zoneId: null == zoneId ? _self.zoneId : zoneId // ignore: cast_nullable_to_non_nullable
-as String,exposes: freezed == exposes ? _self.exposes : exposes // ignore: cast_nullable_to_non_nullable
+as String?,zoneId: freezed == zoneId ? _self.zoneId : zoneId // ignore: cast_nullable_to_non_nullable
+as String?,exposes: freezed == exposes ? _self.exposes : exposes // ignore: cast_nullable_to_non_nullable
 as List<ZigbeeDeviceExpose>?,
   ));
 }
@@ -168,7 +168,7 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String ieeeAddress,  String? manufacturer,  String? model,  DateTime? lastSeen,  String? bindedDeviceId,  String zoneId,  List<ZigbeeDeviceExpose>? exposes)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String ieeeAddress,  String? manufacturer,  String? model, @TimestampOrNullConverter()  DateTime? lastSeen,  String? bindedDeviceId,  String? zoneId,  List<ZigbeeDeviceExpose>? exposes)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _ZigbeeDevice() when $default != null:
 return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.lastSeen,_that.bindedDeviceId,_that.zoneId,_that.exposes);case _:
@@ -189,7 +189,7 @@ return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String ieeeAddress,  String? manufacturer,  String? model,  DateTime? lastSeen,  String? bindedDeviceId,  String zoneId,  List<ZigbeeDeviceExpose>? exposes)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String ieeeAddress,  String? manufacturer,  String? model, @TimestampOrNullConverter()  DateTime? lastSeen,  String? bindedDeviceId,  String? zoneId,  List<ZigbeeDeviceExpose>? exposes)  $default,) {final _that = this;
 switch (_that) {
 case _ZigbeeDevice():
 return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.lastSeen,_that.bindedDeviceId,_that.zoneId,_that.exposes);case _:
@@ -209,7 +209,7 @@ return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String ieeeAddress,  String? manufacturer,  String? model,  DateTime? lastSeen,  String? bindedDeviceId,  String zoneId,  List<ZigbeeDeviceExpose>? exposes)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String ieeeAddress,  String? manufacturer,  String? model, @TimestampOrNullConverter()  DateTime? lastSeen,  String? bindedDeviceId,  String? zoneId,  List<ZigbeeDeviceExpose>? exposes)?  $default,) {final _that = this;
 switch (_that) {
 case _ZigbeeDevice() when $default != null:
 return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.lastSeen,_that.bindedDeviceId,_that.zoneId,_that.exposes);case _:
@@ -224,7 +224,7 @@ return $default(_that.id,_that.ieeeAddress,_that.manufacturer,_that.model,_that.
 @JsonSerializable()
 
 class _ZigbeeDevice extends ZigbeeDevice {
-  const _ZigbeeDevice({required this.id, required this.ieeeAddress, this.manufacturer, this.model, this.lastSeen, this.bindedDeviceId, required this.zoneId, final  List<ZigbeeDeviceExpose>? exposes}): _exposes = exposes,super._();
+  const _ZigbeeDevice({required this.id, required this.ieeeAddress, this.manufacturer, this.model, @TimestampOrNullConverter() this.lastSeen, this.bindedDeviceId, this.zoneId, final  List<ZigbeeDeviceExpose>? exposes}): _exposes = exposes,super._();
   factory _ZigbeeDevice.fromJson(Map<String, dynamic> json) => _$ZigbeeDeviceFromJson(json);
 
 /// [id] is the unique identifier of the device.
@@ -236,11 +236,11 @@ class _ZigbeeDevice extends ZigbeeDevice {
 /// [model] is the model name of the device.
 @override final  String? model;
 /// [lastSeen] is the date and time when the device was last seen.
-@override final  DateTime? lastSeen;
+@override@TimestampOrNullConverter() final  DateTime? lastSeen;
 /// [bindedDeviceId] is the ID of the device that is binded to this device.
 @override final  String? bindedDeviceId;
 /// [zoneId] is the ID of the zone that the device belongs to.
-@override final  String zoneId;
+@override final  String? zoneId;
 /// [exposes] is the list of exposes of the device.
  final  List<ZigbeeDeviceExpose>? _exposes;
 /// [exposes] is the list of exposes of the device.
@@ -286,7 +286,7 @@ abstract mixin class _$ZigbeeDeviceCopyWith<$Res> implements $ZigbeeDeviceCopyWi
   factory _$ZigbeeDeviceCopyWith(_ZigbeeDevice value, $Res Function(_ZigbeeDevice) _then) = __$ZigbeeDeviceCopyWithImpl;
 @override @useResult
 $Res call({
- String id, String ieeeAddress, String? manufacturer, String? model, DateTime? lastSeen, String? bindedDeviceId, String zoneId, List<ZigbeeDeviceExpose>? exposes
+ String id, String ieeeAddress, String? manufacturer, String? model,@TimestampOrNullConverter() DateTime? lastSeen, String? bindedDeviceId, String? zoneId, List<ZigbeeDeviceExpose>? exposes
 });
 
 
@@ -303,7 +303,7 @@ class __$ZigbeeDeviceCopyWithImpl<$Res>
 
 /// Create a copy of ZigbeeDevice
 /// with the given fields replaced by the non-null parameter values.
-@override @pragma('vm:prefer-inline') $Res call({Object? id = null,Object? ieeeAddress = null,Object? manufacturer = freezed,Object? model = freezed,Object? lastSeen = freezed,Object? bindedDeviceId = freezed,Object? zoneId = null,Object? exposes = freezed,}) {
+@override @pragma('vm:prefer-inline') $Res call({Object? id = null,Object? ieeeAddress = null,Object? manufacturer = freezed,Object? model = freezed,Object? lastSeen = freezed,Object? bindedDeviceId = freezed,Object? zoneId = freezed,Object? exposes = freezed,}) {
   return _then(_ZigbeeDevice(
 id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
 as String,ieeeAddress: null == ieeeAddress ? _self.ieeeAddress : ieeeAddress // ignore: cast_nullable_to_non_nullable
@@ -311,8 +311,8 @@ as String,manufacturer: freezed == manufacturer ? _self.manufacturer : manufactu
 as String?,model: freezed == model ? _self.model : model // ignore: cast_nullable_to_non_nullable
 as String?,lastSeen: freezed == lastSeen ? _self.lastSeen : lastSeen // ignore: cast_nullable_to_non_nullable
 as DateTime?,bindedDeviceId: freezed == bindedDeviceId ? _self.bindedDeviceId : bindedDeviceId // ignore: cast_nullable_to_non_nullable
-as String?,zoneId: null == zoneId ? _self.zoneId : zoneId // ignore: cast_nullable_to_non_nullable
-as String,exposes: freezed == exposes ? _self._exposes : exposes // ignore: cast_nullable_to_non_nullable
+as String?,zoneId: freezed == zoneId ? _self.zoneId : zoneId // ignore: cast_nullable_to_non_nullable
+as String?,exposes: freezed == exposes ? _self._exposes : exposes // ignore: cast_nullable_to_non_nullable
 as List<ZigbeeDeviceExpose>?,
   ));
 }
@@ -332,7 +332,7 @@ mixin _$ZigbeeDeviceExpose {
  Map<String, dynamic>? get params;/// [widget] is the rendering widget, useful to render visually the kind of device.
 /// It can be one of the following:
 @JsonKey(unknownEnumValue: RenderWidget.unknown) RenderWidget? get widget;/// [updatedAt] is the date and time when the expose was last updated.
- DateTime? get updatedAt;
+@TimestampOrNullConverter() DateTime? get updatedAt;
 /// Create a copy of ZigbeeDeviceExpose
 /// with the given fields replaced by the non-null parameter values.
 @JsonKey(includeFromJson: false, includeToJson: false)
@@ -365,7 +365,7 @@ abstract mixin class $ZigbeeDeviceExposeCopyWith<$Res>  {
   factory $ZigbeeDeviceExposeCopyWith(ZigbeeDeviceExpose value, $Res Function(ZigbeeDeviceExpose) _then) = _$ZigbeeDeviceExposeCopyWithImpl;
 @useResult
 $Res call({
- String id, String name, String type, String access, Map<String, dynamic>? params,@JsonKey(unknownEnumValue: RenderWidget.unknown) RenderWidget? widget, DateTime? updatedAt
+ String id, String name, String type, String access, Map<String, dynamic>? params,@JsonKey(unknownEnumValue: RenderWidget.unknown) RenderWidget? widget,@TimestampOrNullConverter() DateTime? updatedAt
 });
 
 
@@ -476,7 +476,7 @@ return $default(_that);case _:
 /// }
 /// ```
 
-@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget,  DateTime? updatedAt)?  $default,{required TResult orElse(),}) {final _that = this;
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget, @TimestampOrNullConverter()  DateTime? updatedAt)?  $default,{required TResult orElse(),}) {final _that = this;
 switch (_that) {
 case _ZigbeeDeviceExpose() when $default != null:
 return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.widget,_that.updatedAt);case _:
@@ -497,7 +497,7 @@ return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.w
 /// }
 /// ```
 
-@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget,  DateTime? updatedAt)  $default,) {final _that = this;
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget, @TimestampOrNullConverter()  DateTime? updatedAt)  $default,) {final _that = this;
 switch (_that) {
 case _ZigbeeDeviceExpose():
 return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.widget,_that.updatedAt);case _:
@@ -517,7 +517,7 @@ return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.w
 /// }
 /// ```
 
-@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget,  DateTime? updatedAt)?  $default,) {final _that = this;
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String name,  String type,  String access,  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown)  RenderWidget? widget, @TimestampOrNullConverter()  DateTime? updatedAt)?  $default,) {final _that = this;
 switch (_that) {
 case _ZigbeeDeviceExpose() when $default != null:
 return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.widget,_that.updatedAt);case _:
@@ -532,7 +532,7 @@ return $default(_that.id,_that.name,_that.type,_that.access,_that.params,_that.w
 @JsonSerializable()
 
 class _ZigbeeDeviceExpose extends ZigbeeDeviceExpose {
-  const _ZigbeeDeviceExpose({required this.id, required this.name, required this.type, required this.access, final  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown) this.widget, this.updatedAt}): _params = params,super._();
+  const _ZigbeeDeviceExpose({required this.id, required this.name, required this.type, required this.access, final  Map<String, dynamic>? params, @JsonKey(unknownEnumValue: RenderWidget.unknown) this.widget, @TimestampOrNullConverter() this.updatedAt}): _params = params,super._();
   factory _ZigbeeDeviceExpose.fromJson(Map<String, dynamic> json) => _$ZigbeeDeviceExposeFromJson(json);
 
 /// [id] is the unique identifier of the expose.
@@ -558,7 +558,7 @@ class _ZigbeeDeviceExpose extends ZigbeeDeviceExpose {
 /// It can be one of the following:
 @override@JsonKey(unknownEnumValue: RenderWidget.unknown) final  RenderWidget? widget;
 /// [updatedAt] is the date and time when the expose was last updated.
-@override final  DateTime? updatedAt;
+@override@TimestampOrNullConverter() final  DateTime? updatedAt;
 
 /// Create a copy of ZigbeeDeviceExpose
 /// with the given fields replaced by the non-null parameter values.
@@ -593,7 +593,7 @@ abstract mixin class _$ZigbeeDeviceExposeCopyWith<$Res> implements $ZigbeeDevice
   factory _$ZigbeeDeviceExposeCopyWith(_ZigbeeDeviceExpose value, $Res Function(_ZigbeeDeviceExpose) _then) = __$ZigbeeDeviceExposeCopyWithImpl;
 @override @useResult
 $Res call({
- String id, String name, String type, String access, Map<String, dynamic>? params,@JsonKey(unknownEnumValue: RenderWidget.unknown) RenderWidget? widget, DateTime? updatedAt
+ String id, String name, String type, String access, Map<String, dynamic>? params,@JsonKey(unknownEnumValue: RenderWidget.unknown) RenderWidget? widget,@TimestampOrNullConverter() DateTime? updatedAt
 });