+++ lib/src/zigbee/zigbee.g.dart
@@ -12,11 +12,11 @@ _ZigbeeDevice _$ZigbeeDeviceFromJson(Map<String, dynamic> json) =>
       ieeeAddress: json['ieeeAddress'] as String,
       manufacturer: json['manufacturer'] as String?,
       model: json['model'] as String?,
-      lastSeen: json['lastSeen'] == null
-          ? null
-          : DateTime.parse(json['lastSeen'] as String),
+      lastSeen: const TimestampOrNullConverter().fromJson(
+        json['lastSeen'] as num?,
+      ),
       bindedDeviceId: json['bindedDeviceId'] as String?,
-      zoneId: json['zoneId'] as String,
+      zoneId: json['zoneId'] as String?,
       exposes: (json['exposes'] as List<dynamic>?)
           ?.map((e) => ZigbeeDeviceExpose.fromJson(e as Map<String, dynamic>))
           .toList(),
@@ -28,7 +28,7 @@ Map<String, dynamic> _$ZigbeeDeviceToJson(_ZigbeeDevice instance) =>
       'ieeeAddress': instance.ieeeAddress,
       'manufacturer': instance.manufacturer,
       'model': instance.model,
-      'lastSeen': instance.lastSeen?.toIso8601String(),
+      'lastSeen': const TimestampOrNullConverter().toJson(instance.lastSeen),
       'bindedDeviceId': instance.bindedDeviceId,
       'zoneId': instance.zoneId,
       'exposes': instance.exposes?.map((e) => e.toJson()).toList(),
@@ -46,9 +46,9 @@ _ZigbeeDeviceExpose _$ZigbeeDeviceExposeFromJson(Map<String, dynamic> json) =>
         json['widget'],
         unknownValue: RenderWidget.unknown,
       ),
-      updatedAt: json['updatedAt'] == null
-          ? null
-          : DateTime.parse(json['updatedAt'] as String),
+      updatedAt: const TimestampOrNullConverter().fromJson(
+        json['updatedAt'] as num?,
+      ),
     );
 
 Map<String, dynamic> _$ZigbeeDeviceExposeToJson(_ZigbeeDeviceExpose instance) =>
@@ -59,7 +59,7 @@ Map<String, dynamic> _$ZigbeeDeviceExposeToJson(_ZigbeeDeviceExpose instance) =>
       'access': instance.access,
       'params': instance.params,
       'widget': instance.widget?.toJson(),
-      'updatedAt': instance.updatedAt?.toIso8601String(),
+      'updatedAt': const TimestampOrNullConverter().toJson(instance.updatedAt),
     };
 
 const _$RenderWidgetEnumMap = {